services/certificate-service/requirements.txt
# Django Core
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-filter>=23.0
django-cors-headers>=4.0.0

# Database
psycopg2-binary>=2.9.9

# Authentication
PyJWT>=2.8.0

# Async Support
uvicorn>=0.24.0
gunicorn>=21.0.0

# Validation
pydantic>=2.5.0

# HTTP Client
httpx>=0.25.0
aiohttp>=3.9.0

# Caching
redis>=5.0.0
django-redis>=5.4.0

# Task Queue
celery>=5.3.0

# Monitoring
sentry-sdk>=1.38.0

# Utils
python-dateutil>=2.8.2
pytz>=2023.3

# Testing
pytest>=7.4.0
pytest-django>=4.7.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
factory-boy>=3.3.0

# Code Quality
black>=23.0.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.7.0
